% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_append_versions.R
\name{susor_append_versions}
\alias{susor_append_versions}
\title{Appends all versions of a questionnaire and saves appended data in \code{susor_dir_raw}}
\usage{
susor_append_versions(susor_qn_variable, susor_format = "STATA")
}
\arguments{
\item{susor_qn_variable}{A string. The variable name of questionnaire which versions
are to be appended}

\item{susor_format}{A string. Format of the downloaded data c("STATA", "Tabular").
Default = "STATA"}
}
\value{
A directory with the appended file(s) with all the versions of \code{susor_qn_variable} saved
in \code{susor_dir_raw}.
do not return if file is empty
}
\description{
First define \code{susor_dir_downloads} and \code{susor_dir_raw} in \code{susor_login()} for
\code{susor_append_versions()} to work. Pass the questionnaire variable (\code{susor_qn_variable})
to \code{susor_append_versions()} to append all the versions of that questionnaire and save the
appended data in \code{susor_dir_raw}.
}
