% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_approve_hq.R
\name{susor_approve_hq}
\alias{susor_approve_hq}
\title{Approve interviews as headquarters}
\usage{
susor_approve_hq(
  susor_interview_key,
  susor_approve_comment = NULL,
  susor_data_reference,
  ...
)
}
\arguments{
\item{susor_interview_key}{A string. interview__key of the interview to be a
approved in the server. It can be a vector containing more than one interview__key}

\item{susor_data_reference}{A tibble. A tibble that includes the interviews to
be approved.}

\item{susor_data_comment}{A string. Approve comment}
}
\value{
Change status of interviews to ApprovedByHeadquarters
}
\description{
Approve interviews as headquarters
}
\examples{
#define interview__key of interviews to approve by HQ
to_approve <- c("00-06-27-10","00-78-23-02" )
#approve interviews using 'susor_approve_hq'
susor_approve_hq(susor_interview_key = to_approve,
                by = "interview__key",
                susor_data_reference = raw
                )
}
