% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_generate_file.R
\name{susor_generate_file}
\alias{susor_generate_file}
\title{Start file creation in Survey Solution's server}
\usage{
susor_generate_file(
  susor_quid,
  susor_interview_status = "ALL",
  susor_format = "STATA",
  ...
)
}
\arguments{
\item{susor_quid}{A string. Questionnaire Identity. You can find it in
the tibble \code{susor_questionnaires} that is created by \code{susor_credentials}
See ?\code{susor_credentials}}

\item{susor_interview_status}{A string. To define the status of interviews to download
c( "All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor",
 "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters").
 Default = "ALL"

define API
create .json file temporarly}

\item{susor_format}{A string to define the format of the file c("STATA", "Tabular").
Default = "STATA"}
}
\description{
As for any other function of \code{susor}, \code{susor_generate_file} requires
that you first define your credentials with \code{susor_credentials}.
\code{susor_generate_file} will starts the file creation in the server.
}
