% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_get_assignments.R
\name{susor_get_assignments}
\alias{susor_get_assignments}
\title{Downloads assignments from the server}
\usage{
susor_get_assignments(susor_qn_version, susor_qn_variable, susor_limit = 40)
}
\arguments{
\item{susor_qn_version}{A number with the version of the assignment}

\item{susor_qn_variable}{A string with the variable name of the questionnaire}
}
\value{
A tibbble with the information of the assignment
}
\description{
Downloads assignments from the server
}
