% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_get_diagostics.R
\name{susor_get_diagostics}
\alias{susor_get_diagostics}
\title{Gets interview diagnostics and actions into the main questionnaire}
\usage{
susor_get_diagostics(data_directory = susor_dir_raw, susor_qn_variable)
}
\arguments{
\item{data_directory}{A directory. The directory where the data is stored. This
directory must include the following files: interview__diagnostics.dta,
interview__actions.dta and the questionnaire for which the indicators will be attached}

\item{susor_qn_variable}{The variable name of questionnaire which versions
are to be appended}
}
\description{
Gets interview diagnostics and actions into the main questionnaire
}
