% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_get_interviewers.R
\name{susor_get_interviewers}
\alias{susor_get_interviewers}
\title{Downloads the list of interviewers for a given Supervisor}
\usage{
susor_get_interviewers(
  susor_server,
  susor_user,
  susor_password,
  supervisor = "Supervisor"
)
}
\arguments{
\item{supervisor}{Name of supervisor}
}
\description{
Downloads the list of interviewers for a given Supervisor
}
