% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_get_stata_labels.R
\name{susor_get_stata_labels}
\alias{susor_get_stata_labels}
\title{Assign labels to categorical values using stata attributes}
\usage{
susor_get_stata_labels(x)
}
\description{
Assign labels to categorical values using stata attributes
}
