% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_get_status_creation.R
\name{susor_get_status_creation}
\alias{susor_get_status_creation}
\title{Gets the status of the export process}
\usage{
susor_get_status_creation(apiGenerate, jobID)
}
\arguments{
\item{apiGenerate}{A string. susor_server//api/v2/export}
}
\value{
The status of export process
}
\description{
Gets the status of the export process
}
