% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_login.R
\name{susor_login}
\alias{susor_login}
\title{Define API credentials of your Survey Solutions server}
\usage{
susor_login(
  susor_server,
  susor_user,
  susor_password,
  susor_dir_downloads = "downloads",
  susor_dir_raw = "raw",
  limit = 40,
  ofset = 1,
  ...
)
}
\arguments{
\item{susor_server}{A string. URL of your Survey Solution's server.}

\item{susor_user}{A string. The API user defined in
Survey Solutions Server.}

\item{susor_password}{A string. Password of the API user.}

\item{susor_dir_downloads}{A string. Path to the directory where the
data exported from the server is going to be saved in your local drive.
Default = 'downloads'.}

\item{susor_dir_raw}{A string. Path to the directory where the raw data will
be stored in your local drive.Default = 'raw.}

\item{limit}{A number. Limit number of returned rows. Max allowed
value is 40. Default=40}

\item{offset}{Page number starting from 1. Default = 1}
}
\value{
All the parameters provided to the function (\code{susor_server},
\code{susor_user}, \code{susor_password}, \code{susor_dir_downloads}, and
\code{susor_dir_raw}) will be saved in your global environment so can be used
in your workflow. Additionally, A tibble named \code{susor_questionnaires}
will be saved in your global environment. View this tibble to get: Questionnaire ID,
etc.
}
\description{
\code{susor_login()} checks that your API cdedentials \code{susor_server},
\code{susor_user}, and \code{susor_passowrd} are correct and
saves these credentials in the global environment so they can be used to communicate with the server
This function must be run before any other function of this
package \code{susor}.
}
\examples{
susor_login(susor_server = "http://my.muvasurveys.com",
                  susor_user = "Diego_Maradona",
                  susor_password = "Pelusa123",
                  susor_dir_downloads = file.path("Dropbox", "downloads"),
                  susor_dir_raw = file.path("Dropbox", "raw")
)



}
