% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susor_export_file.R
\name{susor_export_file}
\alias{susor_export_file}
\title{Starts the export process and downloads files from Survey Solution's server.}
\usage{
susor_export_file(
  susor_qn_variable,
  susor_qn_version,
  susor_format = "STATA",
  susor_interview_status = "All",
  temporary_dir = tempdir(),
  ...
)
}
\arguments{
\item{susor_qn_variable}{A string. The variable name of questionnaire (you can
find it in the tibble \code{susor_questionnaires} that is created by
\code{susor_login()}).}

\item{susor_qn_version}{A number. Version number of the questionnaire
to download (it can be found in the tibble \code{susor_questionnaires}
that is created by  \code{susor_login()}).}

\item{susor_format}{A string to define the format of the file c("STATA", "Tabular").
Default = "STATA"}

\item{susor_interview_status}{A string. To define the status of interviews to download
c( "All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor",
 "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters").
 Default = "ALL"}

\item{temporary_dir}{A temporary directory where the .zip format from the
API export response is temporally stored and unzipped from. Default = tempdir().}
}
\value{
Unzipped file saved in \code{susor_dir_downloads}.
}
\description{
As for any other function of \code{susor}, \code{susor_export_file()} requires
that you first define your server credentials with \code{susor_login()}.For
\code{susor_export_file()} it is required that you define \code{susor_dir_downloads}
in \code{susor_login()}. \code{susor_dir_download} is the directory
where the file will be unzipped and saved. Pass the questionnaire variable
\code{susor_qn_variable}, and the version (\code{susor_qn_version}) to download
\code{susor_export_file()} will generate and download
that version of the questionnaire.
}
\examples{
susor_exportfile(susor_qn_variable = "pupils_SLK"
susor_qn_version = 5)
}
